Update parameter generators
