+++ openapi-core/src/main/java/io/ballerina/openapi/core/service/ServiceGenerator.java
@@ -121,143 +121,17 @@ public abstract class ServiceGenerator {
                     if ((operationTags != null && GeneratorUtils.hasTags(operationTags, filterTags)) ||
                             ((operation.getValue().getOperationId() != null) &&
                                     filterOperations.contains(operation.getValue().getOperationId().trim()))) {
-                        // getRelative resource path
-                        List<Node> functionRelativeResourcePath = getRelativeResourcePath(path,
-                                operation.getValue(), oasServiceMetadata.getOpenAPI().getComponents(),
-                                oasServiceMetadata.generateWithoutDataBinding());
                         FunctionDefinitionNode resourceFunction = resourceGenerator
-                                .generateResourceFunction(operation, functionRelativeResourcePath, path);
+                                .generateResourceFunction(operation, path);
                         functions.add(resourceFunction);
                     }
                 }
             } else {
-                // getRelative resource path
-                List<Node> relativeResourcePath = getRelativeResourcePath(path, operation.getValue(),
-                        oasServiceMetadata.getOpenAPI().getComponents(), oasServiceMetadata.generateWithoutDataBinding());
                 FunctionDefinitionNode resourceFunction = resourceGenerator
-                        .generateResourceFunction(operation, relativeResourcePath, path);
+                        .generateResourceFunction(operation, path);
                 functions.add(resourceFunction);
             }
         }
         return functions;
-    }
-
-    /**
-     * Generated resource function relative path node list.
-     *
-     * @param path      - resource path
-     * @param operation - resource operation
-     * @return - node lists
-     * @throws OASTypeGenException
-     */
-    private List<Node> getRelativeResourcePath(String path, Operation operation,
-                                                     Components components, boolean isWithoutDataBinding) {
-
-        List<Node> functionRelativeResourcePath = new ArrayList<>();
-        String[] pathNodes = path.split(GeneratorConstants.SLASH);
-        if (pathNodes.length >= 2) {
-            for (String pathNode : pathNodes) {
-                if (pathNode.contains(GeneratorConstants.OPEN_CURLY_BRACE)) {
-                    String pathParam = pathNode;
-                    pathParam = pathParam.substring(pathParam.indexOf(GeneratorConstants.OPEN_CURLY_BRACE) + 1);
-                    pathParam = pathParam.substring(0, pathParam.indexOf(GeneratorConstants.CLOSE_CURLY_BRACE));
-                    pathParam = getValidName(pathParam, false);
-
-                    /**
-                     * TODO -> `onCall/[string id]\.json` type of url won't support from syntax
-                     * issue https://github.com/ballerina-platform/ballerina-spec/issues/1138
-                     * <pre>resource function get onCall/[string id]\.json() returns string {}</>
-                     */
-                    if (operation.getParameters() != null) {
-                        extractPathParameterDetails(operation, functionRelativeResourcePath, pathNode,
-                                pathParam, components, isWithoutDataBinding);
-                    }
-                } else if (!pathNode.isBlank()) {
-                    IdentifierToken idToken = createIdentifierToken(escapeIdentifier(pathNode.trim()));
-                    functionRelativeResourcePath.add(idToken);
-                    functionRelativeResourcePath.add(createToken(SLASH_TOKEN));
-                }
-            }
-            functionRelativeResourcePath.remove(functionRelativeResourcePath.size() - 1);
-        } else if (pathNodes.length == 0) {
-            IdentifierToken idToken = createIdentifierToken(".");
-            functionRelativeResourcePath.add(idToken);
-        } else {
-            IdentifierToken idToken = createIdentifierToken(pathNodes[1].trim());
-            functionRelativeResourcePath.add(idToken);
-        }
-        return functionRelativeResourcePath;
-    }
-
-
-    private void extractPathParameterDetails(Operation operation, List<Node> functionRelativeResourcePath,
-                                                    String pathNode, String pathParam,
-                                                    Components components, boolean isWithoutDataBinding) {
-        // check whether path parameter segment has special character
-        String[] split = pathNode.split(io.ballerina.openapi.core.generators.type.GeneratorConstants.CLOSE_CURLY_BRACE, 2);
-        Pattern pattern = Pattern.compile(io.ballerina.openapi.core.generators.type.GeneratorConstants.SPECIAL_CHARACTERS_REGEX);
-        Matcher matcher = pattern.matcher(split[1]);
-        boolean hasSpecialCharacter = matcher.find();
-
-        for (Parameter parameter : operation.getParameters()) {
-            if (parameter.get$ref() != null) {
-                try {
-                    parameter = GeneratorMetaData.getInstance().getOpenAPI().getComponents()
-                            .getParameters().get(extractReferenceType(parameter.get$ref()));
-                } catch (OASTypeGenException e) {
-                    throw new RuntimeException(e);
-                }
-            }
-            if (parameter.getIn() == null) {
-                continue;
-            }
-            if (pathParam.trim().equals(getValidName(parameter.getName().trim(), false))
-                    && parameter.getIn().equals("path")) {
-                String paramType;
-                if (parameter.getSchema().get$ref() != null) {
-                    try {
-                        paramType = resolveReferenceType(parameter.getSchema(), components, isWithoutDataBinding,
-                                pathParam);
-                    } catch (OASTypeGenException e) {
-                        throw new RuntimeException(e);
-                    }
-                    Schema<?> schema = GeneratorMetaData.getInstance().getOpenAPI().getComponents().getSchemas().get(paramType);
-                    // todo : update this part
-//                    try {
-//                        TypeHandler.getInstance().generateTypeDescriptorForOASSchema(schema, paramType);
-//                    } catch (OASTypeGenException e) {
-//                        throw new RuntimeException(e);
-//                    }
-                } else {
-                    paramType = getPathParameterType(parameter.getSchema(), pathParam);
-                    if (paramType.endsWith(GeneratorConstants.NILLABLE)) {
-                        throw new RuntimeException("Path parameter value cannot be null.");
-                    }
-                }
-
-                // TypeDescriptor
-                BuiltinSimpleNameReferenceNode builtSNRNode = createBuiltinSimpleNameReferenceNode(
-                        null,
-                        parameter.getSchema() == null || hasSpecialCharacter ?
-                                createIdentifierToken(io.ballerina.openapi.core.generators.type.GeneratorConstants.STRING) :
-                                createIdentifierToken(paramType));
-                IdentifierToken paramName = createIdentifierToken(
-                        hasSpecialCharacter ?
-                                getValidName(pathNode, false) :
-                                pathParam);
-                ResourcePathParameterNode resourcePathParameterNode =
-                        createResourcePathParameterNode(
-                                SyntaxKind.RESOURCE_PATH_SEGMENT_PARAM,
-                                createToken(OPEN_BRACKET_TOKEN),
-                                NodeFactory.createEmptyNodeList(),
-                                builtSNRNode,
-                                null,
-                                paramName,
-                                createToken(CLOSE_BRACKET_TOKEN));
-                functionRelativeResourcePath.add(resourcePathParameterNode);
-                functionRelativeResourcePath.add(createToken(SLASH_TOKEN));
-                break;
-            }
-        }
     }
 }
