+++ openapi-core/src/main/java/io/ballerina/openapi/core/service/parameter/ParameterGenerator.java
@@ -1,16 +1,27 @@
 package io.ballerina.openapi.core.service.parameter;
 
+import io.ballerina.compiler.syntax.tree.Node;
 import io.ballerina.compiler.syntax.tree.ParameterNode;
+import io.ballerina.compiler.syntax.tree.SyntaxKind;
+import io.ballerina.compiler.syntax.tree.Token;
+import io.ballerina.openapi.core.generators.type.exception.OASTypeGenException;
+import io.ballerina.openapi.core.service.GeneratorConstants;
+import io.ballerina.openapi.core.service.ServiceGenerationUtils;
 import io.swagger.v3.oas.models.OpenAPI;
 import io.swagger.v3.oas.models.Operation;
 import io.swagger.v3.oas.models.PathItem;
+import io.swagger.v3.oas.models.parameters.Parameter;
 
+import java.util.List;
 import java.util.Map;
+
+import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createToken;
 
 public abstract class ParameterGenerator {
 
     OpenAPI openAPI;
     boolean isNullableRequired;
+    // todo : check on how to handle nullable in servicegenerator annotation generation
 
-    abstract ParameterNode generateParameterNode(Map.Entry<PathItem.HttpMethod, Operation> operation);
+    public abstract ParameterNode generateParameterNode(Parameter parameter);
 }
