--- openapi-core/src/main/java/io/ballerina/openapi/core/service/parameter/ParametersGeneratorImpl.java
@@ -1,521 +1,0 @@
-/*
- * Copyright (c) 2022, WSO2 LLC. (http://www.wso2.com). All Rights Reserved.
- *
- * WSO2 Inc. licenses this file to you under the Apache License,
- * Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package io.ballerina.openapi.core.service.parameter;
-
-import io.ballerina.compiler.syntax.tree.AbstractNodeFactory;
-import io.ballerina.compiler.syntax.tree.AnnotationNode;
-import io.ballerina.compiler.syntax.tree.BuiltinSimpleNameReferenceNode;
-import io.ballerina.compiler.syntax.tree.DefaultableParameterNode;
-import io.ballerina.compiler.syntax.tree.IdentifierToken;
-import io.ballerina.compiler.syntax.tree.Node;
-import io.ballerina.compiler.syntax.tree.NodeList;
-import io.ballerina.compiler.syntax.tree.OptionalTypeDescriptorNode;
-import io.ballerina.compiler.syntax.tree.ParameterNode;
-import io.ballerina.compiler.syntax.tree.RequiredParameterNode;
-import io.ballerina.compiler.syntax.tree.SyntaxKind;
-import io.ballerina.compiler.syntax.tree.Token;
-import io.ballerina.compiler.syntax.tree.TypeDescriptorNode;
-import io.ballerina.openapi.core.generators.type.GeneratorUtils;
-import io.ballerina.openapi.core.generators.common.TypeHandler;
-import io.ballerina.openapi.core.generators.type.exception.OASTypeGenException;
-import io.ballerina.openapi.core.service.GeneratorConstants;
-import io.ballerina.openapi.core.service.diagnostic.ServiceDiagnosticMessages;
-import io.ballerina.openapi.core.service.ServiceGenerationUtils;
-import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.Operation;
-import io.swagger.v3.oas.models.PathItem;
-import io.swagger.v3.oas.models.media.Content;
-import io.swagger.v3.oas.models.media.MediaType;
-import io.swagger.v3.oas.models.media.Schema;
-import io.swagger.v3.oas.models.parameters.Parameter;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
-import java.util.Locale;
-import java.util.Map;
-
-import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createEmptyNodeList;
-import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createIdentifierToken;
-import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createNodeList;
-import static io.ballerina.compiler.syntax.tree.AbstractNodeFactory.createToken;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createBuiltinSimpleNameReferenceNode;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createDefaultableParameterNode;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createOptionalTypeDescriptorNode;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createRequiredParameterNode;
-import static io.ballerina.compiler.syntax.tree.NodeFactory.createSimpleNameReferenceNode;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.CLOSE_PAREN_TOKEN;
-import static io.ballerina.compiler.syntax.tree.SyntaxKind.OPEN_PAREN_TOKEN;
-import static io.ballerina.openapi.core.service.GeneratorConstants.NILLABLE;
-import static io.ballerina.openapi.core.service.diagnostic.ServiceDiagnosticMessages.OAS_SERVICE_103;
-import static io.ballerina.openapi.core.service.diagnostic.ServiceDiagnosticMessages.OAS_SERVICE_104;
-import static io.ballerina.openapi.core.service.diagnostic.ServiceDiagnosticMessages.OAS_SERVICE_105;
-import static io.ballerina.openapi.core.service.diagnostic.ServiceDiagnosticMessages.OAS_SERVICE_106;
-import static io.ballerina.openapi.core.generators.type.GeneratorUtils.convertOpenAPITypeToBallerina;
-import static io.ballerina.openapi.core.generators.type.GeneratorUtils.getOpenAPIType;
-import static io.ballerina.openapi.core.generators.type.GeneratorUtils.getValidName;
-import static io.ballerina.openapi.core.generators.type.GeneratorUtils.isArraySchema;
-import static io.ballerina.openapi.core.generators.type.GeneratorUtils.isMapSchema;
-import static io.ballerina.openapi.core.generators.type.GeneratorUtils.isObjectSchema;
-
-/**
- * This class uses for generating all resource function parameters.
- *
- * @since 1.3.0
- */
-public class ParametersGeneratorImpl extends ParameterGenerator {
-
-    private final List<Node> requiredParams;
-    private final List<Node> defaultableParams;
-
-    private static final List<String> paramSupportedTypes =
-            new ArrayList<>(Arrays.asList(GeneratorConstants.INTEGER, GeneratorConstants.NUMBER,
-                    GeneratorConstants.STRING, GeneratorConstants.BOOLEAN));
-
-    public ParametersGeneratorImpl(boolean isNullableRequired, OpenAPI openAPI) {
-        this.isNullableRequired = isNullableRequired;
-        this.openAPI = openAPI;
-        this.requiredParams = new ArrayList<>();
-        this.defaultableParams = new ArrayList<>();
-    }
-
-    @Override
-    ParameterNode generateParameterNode(Map.Entry<PathItem.HttpMethod, Operation> operation) {
-        return null;
-    }
-
-    public List<Node> getRequiredParams() {
-        return requiredParams;
-    }
-
-    public List<Node> getDefaultableParams() {
-        return defaultableParams;
-    }
-
-    public boolean isNullableRequired() {
-        return isNullableRequired;
-    }
-
-    /**
-     * This function for generating operation parameters.
-     *
-     * @param operation OAS operation
-     * @throws OASTypeGenException when the parameter generation fails.
-     */
-    public void generateResourcesInputs(Map.Entry<PathItem.HttpMethod, Operation> operation)
-            throws OASTypeGenException {
-
-        Token comma = createToken(SyntaxKind.COMMA_TOKEN);
-        // Handle header and query parameters
-        if (operation.getValue().getParameters() != null) {
-            List<Parameter> parameters = operation.getValue().getParameters();
-            for (Parameter parameter : parameters) {
-                Node param = null;
-                if (parameter.get$ref() != null) {
-                    String referenceType = ServiceGenerationUtils.extractReferenceType(parameter.get$ref());
-                    parameter = openAPI.getComponents().getParameters().get(referenceType);
-                }
-                if (parameter.getIn().trim().equals(GeneratorConstants.HEADER)) {
-                    param = handleHeader(parameter);
-                    if (param.kind() == SyntaxKind.DEFAULTABLE_PARAM) {
-                        defaultableParams.add(param);
-                        defaultableParams.add(comma);
-                    } else {
-                        requiredParams.add(param);
-                        requiredParams.add(comma);
-                    }
-                } else if (parameter.getIn().trim().equals(GeneratorConstants.QUERY)) {
-                    if (parameter.getRequired() != null && parameter.getRequired() &&
-                            (parameter.getSchema() != null && parameter.getSchema().getNullable() != null &&
-                                    parameter.getSchema().getNullable())) {
-                        isNullableRequired = true;
-                    }
-                    // type  BasicType boolean|int|float|decimal|string ;
-                    // public type () |BasicType|BasicType []| map<json>;
-                    param = createNodeForQueryParam(parameter);
-                    if (param != null) {
-                        if (param.kind() == SyntaxKind.DEFAULTABLE_PARAM) {
-                            defaultableParams.add(param);
-                            defaultableParams.add(comma);
-                        } else {
-                            requiredParams.add(param);
-                            requiredParams.add(comma);
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    /**
-     * This function for generating parameter ST node for header.
-     * <pre> resource function get pets(@http:Header {name:"x-request-id"} string header) </pre>
-     */
-    private ParameterNode handleHeader(Parameter parameter) throws OASTypeGenException {
-
-        Schema<?> schema = parameter.getSchema();
-        String headerType;
-        TypeDescriptorNode headerTypeName;
-        IdentifierToken parameterName = createIdentifierToken(GeneratorUtils.escapeIdentifier(parameter.getName()
-                        .toLowerCase(Locale.ENGLISH)), AbstractNodeFactory.createEmptyMinutiaeList(),
-                GeneratorUtils.SINGLE_WS_MINUTIAE);
-
-        if (getOpenAPIType(schema) == null && schema.get$ref() == null) {
-            // Header example:
-            // 01.<pre>
-            //       in: header
-            //       name: X-Request-ID
-            //       schema: {}
-            //  </pre>
-            throw new OASTypeGenException(String.format(OAS_SERVICE_106.getDescription(), parameter.getName()));
-        } else if (schema.get$ref() != null) {
-            String type = getValidName(ServiceGenerationUtils.extractReferenceType(schema.get$ref()), true);
-            Schema<?> refSchema = openAPI.getComponents().getSchemas().get(type.trim());
-            if (paramSupportedTypes.contains(getOpenAPIType(refSchema)) || isArraySchema(refSchema)) {
-                headerType = type;
-            } else {
-                throw new OASTypeGenException(String.format(OAS_SERVICE_105.getDescription(),
-                        parameter.getName(), getOpenAPIType(refSchema)));
-            }
-        } else if (paramSupportedTypes.contains(getOpenAPIType(schema)) || isArraySchema(schema)) {
-            headerType = convertOpenAPITypeToBallerina(schema).trim();
-        } else {
-            throw new OASTypeGenException(String.format(OAS_SERVICE_105.getDescription(),
-                    parameter.getName(), getOpenAPIType(schema)));
-        }
-
-        if (isArraySchema(schema)) {
-            // TODO: Support nested arrays
-            Schema<?> items = schema.getItems();
-            String arrayType;
-            if (getOpenAPIType(items) == null && items.get$ref() == null) {
-                throw new OASTypeGenException(String.format(OAS_SERVICE_104.getDescription(),
-                        parameter.getName()));
-            } else if (items.get$ref() != null) {
-                String type = getValidName(ServiceGenerationUtils.extractReferenceType(items.get$ref()), true);
-                Schema<?> refSchema = openAPI.getComponents().getSchemas().get(type.trim());
-                if (paramSupportedTypes.contains(getOpenAPIType(refSchema))) {
-                    arrayType = type;
-                } else {
-                    throw new OASTypeGenException(String.format(OAS_SERVICE_103.getDescription(),
-                            parameter.getName(), type));
-                }
-            } else if (!paramSupportedTypes.contains(getOpenAPIType(items))) {
-                throw new OASTypeGenException(String.format(OAS_SERVICE_103.getDescription(),
-                        parameter.getName(), getOpenAPIType(items)));
-            } else if (items.getEnum() != null && !items.getEnum().isEmpty()) {
-                arrayType = OPEN_PAREN_TOKEN.stringValue() + convertOpenAPITypeToBallerina(items) +
-                        CLOSE_PAREN_TOKEN.stringValue();
-            } else {
-                arrayType = convertOpenAPITypeToBallerina(items);
-            }
-            BuiltinSimpleNameReferenceNode headerArrayItemTypeName = createBuiltinSimpleNameReferenceNode(
-                    null, createIdentifierToken(arrayType));
-            // todo : update this part
-            headerTypeName = null;
-//            headerTypeName = TypeHandler.getInstance().getArrayTypeDescriptorNodeFromTypeDescriptorNode(headerArrayItemTypeName);
-        } else {
-            headerTypeName = createBuiltinSimpleNameReferenceNode(null, createIdentifierToken(
-                    headerType, GeneratorUtils.SINGLE_WS_MINUTIAE,
-                    GeneratorUtils.SINGLE_WS_MINUTIAE));
-        }
-        // Create annotation for header
-        // TODO: This code block is to be enabled when handle the headers handle additional parameters
-        // MappingConstructorExpressionNode annotValue = NodeFactory.createMappingConstructorExpressionNode(
-        //        createToken(SyntaxKind.OPEN_BRACE_TOKEN), NodeFactory.createSeparatedNodeList(),
-        //        createToken(SyntaxKind.CLOSE_BRACE_TOKEN));
-
-        AnnotationNode headerNode = ServiceGenerationUtils.getAnnotationNode(GeneratorConstants.HEADER_ANNOT, null);
-        NodeList<AnnotationNode> headerAnnotations = createNodeList(headerNode);
-        // Handle optional values in headers
-        if (!parameter.getRequired()) {
-            // If optional it behaves like default value with null ex:(string? header)
-            // If schema has an enum and that has a null values then the type is already nill. Hence, the check.
-            headerTypeName = headerTypeName.toString().trim().endsWith(NILLABLE) ? headerTypeName :
-                    createOptionalTypeDescriptorNode(headerTypeName,
-                            createToken(SyntaxKind.QUESTION_MARK_TOKEN));
-        }
-        // Handle default values in headers
-        if (schema.getDefault() != null) {
-            return getDefaultableHeaderNode(schema, headerTypeName, parameterName, headerAnnotations);
-        }
-        // Handle header with parameter required true and nullable ture ex: (string? header)
-        if (parameter.getRequired() && schema.getNullable() != null && schema.getNullable().equals(true)) {
-            isNullableRequired = true;
-            headerTypeName = headerTypeName.toString().trim().endsWith(NILLABLE) ? headerTypeName :
-                    createOptionalTypeDescriptorNode(headerTypeName,
-                            createToken(SyntaxKind.QUESTION_MARK_TOKEN));
-        }
-        return createRequiredParameterNode(headerAnnotations, headerTypeName, parameterName);
-    }
-
-    /**
-     * Generate ballerina default headers.
-     */
-    private DefaultableParameterNode getDefaultableHeaderNode(Schema<?> schema, TypeDescriptorNode headerTypeName,
-                                                              IdentifierToken parameterName,
-                                                              NodeList<AnnotationNode> headerAnnotations) {
-
-        if (!getOpenAPIType(schema).equals(GeneratorConstants.ARRAY)) {
-            return createDefaultableParameterNode(headerAnnotations, headerTypeName, parameterName,
-                    createToken(SyntaxKind.EQUAL_TOKEN),
-                    createSimpleNameReferenceNode(createIdentifierToken("\"" +
-                            schema.getDefault().toString() + "\"")));
-        }
-        return createDefaultableParameterNode(headerAnnotations, headerTypeName, parameterName,
-                createToken(SyntaxKind.EQUAL_TOKEN),
-                createSimpleNameReferenceNode(createIdentifierToken(schema.getDefault().toString())));
-    }
-
-    /**
-     * This for generate query parameter nodes.
-     * <p>
-     * Ballerina support query parameter types :
-     * type BasicType boolean|int|float|decimal|string ;
-     * public type  QueryParamType <map>json | () |BasicType|BasicType[];
-     */
-    private Node createNodeForQueryParam(Parameter parameter) throws OASTypeGenException {
-        Schema<?> schema = parameter.getSchema();
-        NodeList<AnnotationNode> annotations = createEmptyNodeList();
-        IdentifierToken parameterName = createIdentifierToken(
-                GeneratorUtils.escapeIdentifier(parameter.getName().trim()),
-                AbstractNodeFactory.createEmptyMinutiaeList(), GeneratorUtils.SINGLE_WS_MINUTIAE);
-        boolean isSchemaNotSupported = schema == null || getOpenAPIType(schema) == null;
-        //Todo: will enable when header parameter support objects
-        //paramSupportedTypes.add(GeneratorConstants.OBJECT);
-        if (schema != null && schema.get$ref() != null) {
-            String type = getValidName(ServiceGenerationUtils.extractReferenceType(schema.get$ref()), true);
-            Schema<?> refSchema = openAPI.getComponents().getSchemas().get(type);
-            return handleReferencedQueryParameter(parameter, type, refSchema, annotations, parameterName);
-        } else if (parameter.getContent() != null) {
-            Content content = parameter.getContent();
-            for (Map.Entry<String, MediaType> mediaTypeEntry : content.entrySet()) {
-                return handleMapJsonQueryParameter(parameter, annotations, parameterName, mediaTypeEntry);
-            }
-        } else if (isSchemaNotSupported) {
-            ServiceDiagnosticMessages messages = ServiceDiagnosticMessages.OAS_SERVICE_102;
-            throw new OASTypeGenException(String.format(messages.getDescription(),
-                    getOpenAPIType(parameter.getSchema())));
-        } else if (parameter.getSchema().getDefault() != null) {
-            // When query parameter has default value
-            return handleDefaultQueryParameter(schema, annotations, parameterName);
-        } else if (parameter.getRequired() && schema.getNullable() == null) {
-            // Required typeDescriptor
-            return handleRequiredQueryParameter(schema, annotations, parameterName);
-        } else {
-            // Optional typeDescriptor
-            return handleOptionalQueryParameter(schema, annotations, parameterName);
-        }
-        return null;
-    }
-
-    /**
-     * Handle query parameter for content type which has application/json.
-     * example:
-     * <pre>
-     *     parameters:
-     *         - name: petType
-     *           in: query
-     *           content:
-     *             application/json:
-     *               schema:
-     *                 type: object
-     *                 additionalProperties: true
-     * </pre>
-     */
-    private RequiredParameterNode handleMapJsonQueryParameter(Parameter parameter, NodeList<AnnotationNode> annotations,
-                                                              IdentifierToken parameterName,
-                                                              Map.Entry<String, MediaType> mediaTypeEntry)
-            throws OASTypeGenException {
-
-        Schema<?> parameterSchema;
-        if (mediaTypeEntry.getValue().getSchema() != null && mediaTypeEntry.getValue().getSchema().get$ref() != null) {
-            String type = getValidName(ServiceGenerationUtils.extractReferenceType(mediaTypeEntry.getValue()
-                    .getSchema().get$ref()), true);
-            parameterSchema = (Schema<?>) openAPI.getComponents().getSchemas().get(type.trim());
-        } else {
-            parameterSchema = mediaTypeEntry.getValue().getSchema();
-        }
-        if (mediaTypeEntry.getKey().equals(GeneratorConstants.APPLICATION_JSON) && isMapSchema(parameterSchema)) {
-            return getMapJsonParameterNode(parameterName, parameter, annotations);
-        }
-        String type = GeneratorUtils.getBallerinaMediaType(mediaTypeEntry.getKey(), false);
-        throw new OASTypeGenException(String.format(ServiceDiagnosticMessages.OAS_SERVICE_102.getDescription(),
-                type));
-    }
-
-    private RequiredParameterNode getMapJsonParameterNode(IdentifierToken parameterName, Parameter parameter,
-                                                         NodeList<AnnotationNode> annotations) {
-        BuiltinSimpleNameReferenceNode rTypeName = createBuiltinSimpleNameReferenceNode(null,
-                createIdentifierToken(io.ballerina.openapi.core.generators.type.GeneratorConstants.MAP_JSON));
-        if (parameter.getRequired()) {
-            return createRequiredParameterNode(annotations, rTypeName, parameterName);
-        }
-        OptionalTypeDescriptorNode optionalNode = createOptionalTypeDescriptorNode(rTypeName,
-                createToken(SyntaxKind.QUESTION_MARK_TOKEN));
-        return createRequiredParameterNode(annotations, optionalNode, parameterName);
-    }
-
-    /**
-     * This function is to handle query schema which does not have required as true.
-     */
-    private Node handleOptionalQueryParameter(Schema<?> schema, NodeList<AnnotationNode> annotations,
-                                              IdentifierToken parameterName) throws OASTypeGenException {
-        if (isArraySchema(schema)) {
-            Schema<?> items = schema.getItems();
-            if (getOpenAPIType(items) == null && items.get$ref() == null) {
-                // Resource function doesn't support to query parameters with array type which doesn't have an
-                // item type.
-                ServiceDiagnosticMessages messages = ServiceDiagnosticMessages.OAS_SERVICE_101;
-                throw new OASTypeGenException(messages.getDescription());
-            } else if ((!(isObjectSchema(items)) && !(getOpenAPIType(items) != null &&
-                    getOpenAPIType(items).equals(GeneratorConstants.ARRAY))) || items.get$ref() != null) {
-
-                // todo : update this part
-                TypeDescriptorNode arrayTypeName = null;
-//                TypeDescriptorNode arrayTypeName = TypeHandler.getInstance().getArrayTypeDescriptorNode(items);
-                OptionalTypeDescriptorNode optionalNode = createOptionalTypeDescriptorNode(arrayTypeName,
-                        createToken(SyntaxKind.QUESTION_MARK_TOKEN));
-                return createRequiredParameterNode(annotations, optionalNode, parameterName);
-            } else if (getOpenAPIType(items).equals(GeneratorConstants.ARRAY)) {
-                // Resource function doesn't support to the nested array type query parameters.
-                ServiceDiagnosticMessages messages = ServiceDiagnosticMessages.OAS_SERVICE_100;
-                throw new OASTypeGenException(messages.getDescription());
-            } else {
-                ServiceDiagnosticMessages messages = ServiceDiagnosticMessages.OAS_SERVICE_102;
-                throw new OASTypeGenException(String.format(messages.getDescription(), "object"));
-            }
-        } else {
-            // todo : update this part
-            Token name = null;
-//            Token name = TypeHandler.getInstance().getQueryParamTypeToken(schema);
-            TypeDescriptorNode queryParamType = createBuiltinSimpleNameReferenceNode(null, name);
-            // If schema has an enum with null value, the type is already nil. Hence, the check.
-            if (!name.text().trim().endsWith(NILLABLE)) {
-                queryParamType = createOptionalTypeDescriptorNode(queryParamType,
-                        createToken(SyntaxKind.QUESTION_MARK_TOKEN));
-            }
-            return createRequiredParameterNode(annotations, queryParamType, parameterName);
-        }
-    }
-
-    private Node handleReferencedQueryParameter(Parameter parameter, String refTypeName, Schema<?> refSchema,
-                                                NodeList<AnnotationNode> annotations, IdentifierToken parameterName) throws OASTypeGenException {
-        // todo : update this part
-        TypeDescriptorNode refTypeNameNode = null;
-//        TypeDescriptorNode refTypeNameNode = TypeHandler.getInstance().getReferencedQueryParameterTypeFromSchema(refSchema, refTypeName);
-        if (refSchema.getDefault() != null) {
-            String defaultValue = getOpenAPIType(refSchema).equals(GeneratorConstants.STRING) ?
-                    String.format("\"%s\"", refSchema.getDefault().toString()) : refSchema.getDefault().toString();
-            return createDefaultableParameterNode(annotations, refTypeNameNode, parameterName,
-                    createToken(SyntaxKind.EQUAL_TOKEN),
-                    createSimpleNameReferenceNode(createIdentifierToken(defaultValue)));
-        } else if (parameter.getRequired() && (refSchema.getNullable() == null || (!refSchema.getNullable()))) {
-            return createRequiredParameterNode(annotations, refTypeNameNode, parameterName);
-        } else {
-            OptionalTypeDescriptorNode optionalNode = createOptionalTypeDescriptorNode(refTypeNameNode,
-                    createToken(SyntaxKind.QUESTION_MARK_TOKEN));
-            return createRequiredParameterNode(annotations, optionalNode, parameterName);
-        }
-    }
-
-    private Node handleRequiredQueryParameter(Schema<?> schema, NodeList<AnnotationNode> annotations,
-                                              IdentifierToken parameterName) throws OASTypeGenException {
-
-        if (isArraySchema(schema)) {
-            Schema<?> items = schema.getItems();
-            if (!(isArraySchema(items)) && (getOpenAPIType(items) != null || (items.get$ref() != null))) {
-                // todo : update this part
-                TypeDescriptorNode arrayTypeName = null;
-//                TypeDescriptorNode arrayTypeName = TypeHandler.getInstance().getArrayTypeDescriptorNode(items);
-                return createRequiredParameterNode(annotations, arrayTypeName, parameterName);
-            } else if (getOpenAPIType(items) == null) {
-                // Resource function doesn't support query parameters for array types that doesn't have an item type.
-                ServiceDiagnosticMessages messages = ServiceDiagnosticMessages.OAS_SERVICE_101;
-                throw new OASTypeGenException(messages.getDescription());
-            } else if (isObjectSchema(items)) {
-                ServiceDiagnosticMessages messages = ServiceDiagnosticMessages.OAS_SERVICE_102;
-                throw new OASTypeGenException(String.format(messages.getDescription(), "object"));
-            } else {
-                // Resource function doesn't support to the nested array type query parameters.
-                ServiceDiagnosticMessages messages = ServiceDiagnosticMessages.OAS_SERVICE_100;
-                throw new OASTypeGenException(messages.getDescription());
-            }
-        } else {
-            // todo : update this part
-            Token name = null;
-//            Token name = TypeHandler.getInstance().getQueryParamTypeToken(schema);
-            BuiltinSimpleNameReferenceNode rTypeName = createBuiltinSimpleNameReferenceNode(null, name);
-            return createRequiredParameterNode(annotations, rTypeName, parameterName);
-        }
-    }
-
-    /**
-     * This function generate default query parameter when OAS gives default value.
-     * <p>
-     * OAS code example:
-     * <pre>
-     *      parameters:
-     *         - name: limit
-     *           in: query
-     *           schema:
-     *             type: integer
-     *             default: 10
-     *             format: int32
-     *  </pre>
-     * generated ballerina -> int limit = 10;
-     */
-
-    private Node handleDefaultQueryParameter(Schema<?> schema, NodeList<AnnotationNode> annotations,
-                                             IdentifierToken parameterName) throws OASTypeGenException {
-
-        if (isArraySchema(schema)) {
-            Schema<?> items = schema.getItems();
-            if (!isArraySchema(items) && (getOpenAPIType(items) != null || (items.get$ref() != null))) {
-                // todo : update this part
-                TypeDescriptorNode arrayTypeName = null;
-//                TypeDescriptorNode arrayTypeName = TypeHandler.getInstance().getArrayTypeDescriptorNode(items);
-                return createDefaultableParameterNode(annotations, arrayTypeName, parameterName,
-                        createToken(SyntaxKind.EQUAL_TOKEN),
-                        createSimpleNameReferenceNode(createIdentifierToken(schema.getDefault().toString())));
-            } else if (getOpenAPIType(items) == null) {
-                // Resource function doesn't support to query parameters with array type which hasn't item type.
-                ServiceDiagnosticMessages messages = ServiceDiagnosticMessages.OAS_SERVICE_101;
-                throw new OASTypeGenException(messages.getDescription());
-            } else {
-                // Resource function doesn't support to the nested array type query parameters.
-                ServiceDiagnosticMessages messages = ServiceDiagnosticMessages.OAS_SERVICE_100;
-                throw new OASTypeGenException(messages.getDescription());
-            }
-        } else {
-            // todo : update this part
-            Token name = null;
-//            Token name = TypeHandler.getInstance().getQueryParamTypeToken(schema);
-            BuiltinSimpleNameReferenceNode rTypeName = createBuiltinSimpleNameReferenceNode(null, name);
-            if (getOpenAPIType(schema).equals(GeneratorConstants.STRING)) {
-                return createDefaultableParameterNode(annotations, rTypeName, parameterName,
-                        createToken(SyntaxKind.EQUAL_TOKEN),
-                        createSimpleNameReferenceNode(createIdentifierToken('"' +
-                                schema.getDefault().toString() + '"')));
-            }
-            return createDefaultableParameterNode(annotations, rTypeName, parameterName,
-                    createToken(SyntaxKind.EQUAL_TOKEN),
-                    createSimpleNameReferenceNode(createIdentifierToken(schema.getDefault().toString())));
-        }
-    }
-}
